Make maven terminal static.
