+++ src/extension.ts
@@ -12,6 +12,7 @@ import { ITreeItem } from "./explorer/model/ITreeItem";
 import { MavenProject } from "./explorer/model/MavenProject";
 import { PluginGoal } from "./explorer/model/PluginGoal";
 import { pluginInfoProvider } from "./explorer/pluginInfoProvider";
+import { mavenTerminal } from "./mavenTerminal";
 import { Settings } from "./Settings";
 import { Utils } from "./Utils";
 import { VSCodeUI } from "./VSCodeUI";
@@ -127,18 +128,18 @@ async function doActivate(_operationId: string, context: vscode.ExtensionContext
     });
 
     context.subscriptions.push(vscode.window.onDidCloseTerminal((closedTerminal: vscode.Terminal) => {
-        VSCodeUI.mavenTerminal.onDidCloseTerminal(closedTerminal);
+        mavenTerminal.onDidCloseTerminal(closedTerminal);
     }));
 
     // configuration change listener
     vscode.workspace.onDidChangeConfiguration((e: vscode.ConfigurationChangeEvent) => {
         // close all terminals with outdated JAVA related Envs
         if (e.affectsConfiguration("maven.terminal.useJavaHome") || e.affectsConfiguration("maven.terminal.customEnv")) {
-            VSCodeUI.mavenTerminal.closeAllTerminals();
+            mavenTerminal.closeAllTerminals();
         } else {
             const useJavaHome: boolean = Settings.Terminal.useJavaHome();
             if (useJavaHome && e.affectsConfiguration("java.home")) {
-                VSCodeUI.mavenTerminal.closeAllTerminals();
+                mavenTerminal.closeAllTerminals();
             }
         }
     });
